requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
charset-normalizer==2.1.1
discord.py==2.0.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
yarl==1.8.1
